rename edit_loops.py -> multi_loops.py
