+++ mmagic/engine/runner/__init__.py
@@ -1,5 +1,5 @@
 # Copyright (c) OpenMMLab. All rights reserved.
-from .edit_loops import MultiTestLoop, MultiValLoop
 from .log_processor import LogProcessor
+from .multi_loops import MultiTestLoop, MultiValLoop
 
 __all__ = ['MultiTestLoop', 'MultiValLoop', 'LogProcessor']
